refactor: updating helper tool configuration

+++ modules/installing/scripts/uninstall.cmake
@@ -15,32 +15,47 @@ cmake_minimum_required (VERSION 3.22 FATAL_ERROR)
 set (files_to_uninstall "")
 
 file (GLOB manifest_files RELATIVE "@CMAKE_BINARY_DIR@"
-	  @CMAKE_BINARY_DIR@/install_manifest_*.txt) # @CMAKE_BINARY_DIR@
+	  @CMAKE_BINARY_DIR@/install_manifest_*.txt) # CMAKE_BINARY_DIR
 
-foreach (manifest_file install_manifest.txt ${manifest_files})
-	set (file_path "@CMAKE_BINARY_DIR@/${manifest_file}")
+# cmake-format: off
+foreach (manifest_file IN LISTS manifest_files
+			ITEMS install_manifest.txt
+			@ORANGES_ADDITIONAL_INSTALL_MANIFEST_FILES@ # ORANGES_ADDITIONAL_INSTALL_MANIFEST_FILES expanded
+	)
+# cmake-format: on
+
+	if (IS_ABSOLUTE "${manifest_file}")
+		set (file_path "${manifest_file}")
+	else ()
+		set (file_path "@CMAKE_BINARY_DIR@/${manifest_file}")
+	endif ()
 
 	if (NOT EXISTS "${file_path}")
 		message (WARNING "Install manifest file ${file_path} does not exist!")
-	else ()
-		message (
-			STATUS "Removing files listed in install manifest: ${file_path}")
-
-		file (STRINGS "${file_path}" installed_files)
-
-		foreach (file ${installed_files})
-			if (IS_SYMLINK "${file}" OR EXISTS "${file}")
-				list (APPEND files_to_uninstall "${file}")
-			else ()
-				message (WARNING "File ${file} does not exist.")
-			endif ()
-		endforeach ()
+		continue ()
 	endif ()
+
+	message (STATUS "Removing files listed in install manifest: ${file_path}")
+
+	file (STRINGS "${file_path}" installed_files)
+
+	foreach (file IN LISTS installed_files)
+		if (IS_SYMLINK "${file}" OR EXISTS "${file}")
+			list (APPEND files_to_uninstall "${file}")
+		else ()
+			message (WARNING "File ${file} does not exist.")
+		endif ()
+	endforeach ()
 endforeach ()
+
+if (NOT files_to_uninstall)
+	message (WARNING "No files found to uninstall.")
+	return ()
+endif ()
 
 list (REMOVE_DUPLICATES files_to_uninstall)
 
-foreach (file ${files_to_uninstall})
+foreach (file IN LISTS files_to_uninstall)
 	message (STATUS "Uninstalling ${file}")
 
 	file (REMOVE "${file}")
